Moved raster-specific mappers to raster_mappers.py.
